requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv==0.19.0
requests==2.26.0

https://www.linkedin.com/in/aaron-mubiana-a3548b2b2/